requirements.txt
dbus-python>=1.3.2
pydbus>=0.6.0
python-dotenv>=1.0.0
SQLAlchemy>=2.0.0
pytest>=7.4.0
black>=23.0.0
pylint>=3.0.0 